requirements.txt
icalendar
requests
google-generativeai